requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.1
fastapi-utils==0.2.1
h11==0.14.0
idna==3.4
pydantic==1.10.12
pydantic_core==2.6.3
python-decouple==3.8
requests==2.31.0
sniffio==1.3.0
SQLAlchemy==1.4.49
starlette==0.27.0
stripe==6.4.0
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
